Combat is functional!
