+++ System Miami/Assets/_Project/_Scripts/_Combat/Helpers/Targets.cs
@@ -14,6 +14,8 @@ namespace SystemMiami
         public List<OverlayTile> Tiles { get { return _tiles; } }
         public List<Combatant> Combatants {  get { return _combatants; } }
 
+        public static readonly Targets empty = new Targets(new List<Vector2Int>(), new List<OverlayTile>(), new List<Combatant>());
+
         public Targets(List<Vector2Int> positions, List<OverlayTile> tiles, List<Combatant> combatants)
         {
             _positions = positions;
