+++ System Miami/Assets/_Project/_Scripts/_Combat/TurnManager.cs
@@ -1,7 +1,7 @@
+// Author: Lee St Louis, Layla Hoey
 using System;
 using System.Collections;
 using System.Collections.Generic;
-using SystemMiami.AbilitySystem;
 using SystemMiami.CombatSystem;
 using SystemMiami.Management;
 using UnityEngine;
@@ -22,6 +22,7 @@ namespace SystemMiami
     /// Manages turns and phases in the combat system.
     /// Handles switching between player and enemy turns,
     /// as well as movement and action phases.
+    /// Spawns enemies.
     /// </summary>
     public class TurnManager : Singleton<TurnManager>
     {
@@ -42,13 +43,17 @@ namespace SystemMiami
         {
             get
             {
+                if (CurrentTurnOwner == null)
+                    { return false; }
+
+                if (CurrentTurnOwner.Controller == null)
+                    { return false; }
+
                 return CurrentTurnOwner.Controller is PlayerController;
             }
         }
 
-        public Action<Combatant> BeginTurn;
         public Action<Phase> NewTurnPhase;
-        public Action<Combatant> EndTurn;
 
         public Combatant CurrentTurnOwner { get; private set; }
 
@@ -63,9 +68,6 @@ namespace SystemMiami
 
             combatants.Add(playerCharacter);
             combatants.AddRange(enemyCharacters);
-
-            //// Initialize turns, start with first player
-            //StartPlayerTurn();
 
             StartCoroutine(TurnSequence());
         }
@@ -88,32 +90,6 @@ namespace SystemMiami
         #region Turn Management
         //===============================
 
-        ///// <summary>
-        ///// Starts the player's turn.
-        ///// Resets movement points and action flags for each player character.
-        ///// </summary>
-        //public void StartPlayerTurn()
-        //{
-        //    playerCharacter.Controller.StartTurn();
-        //    CurrentTurnOwner = playerCharacter;
-
-        //    // Actions for other scripts to use
-        //    BeginTurn?.Invoke(playerCharacter);
-        //    NewTurnPhase?.Invoke(Phase.Movement);
-
-        //    Debug.Log("Player's turn started. Movement Phase.");
-        //}
-
-        ///// <summary>
-        ///// Starts the enemy's turn.
-        ///// Resets movement points and action flags for each enemy character.
-        ///// </summary>
-        //public void StartAllEnemyTurns()
-        //{
-        //    // Start enemy AI coroutine
-        //    StartCoroutine(TurnSequence());
-        //}
-
         /// <summary>
         /// Coroutine for handling enemy turns.
         /// Each enemy takes their movement and action phases in sequence.
@@ -122,6 +98,12 @@ namespace SystemMiami
         {
             while (!IsGameOver)
             {
+                // TODO =====================================================
+                // This doesn't work properly when a combatant is removed
+                // from the list on dying. Not sure if theres an extra step
+                // to the garbage collection beyond just removing it from the
+                // list, but it throws an error when anything dies.
+                // ==========================================================
                 foreach (Combatant combatant in combatants)
                 {
                     if (combatant == null)
@@ -138,34 +120,16 @@ namespace SystemMiami
 
                     yield return new WaitForEndOfFrame();
                     yield return new WaitUntil(() => !combatant.Controller.IsMyTurn);
-
-                    //yield return StartCoroutine(controller.TakeTurn()); 
                 }
 
                 yield return null;
             }
         }
 
-
-        ///// <summary>
-        ///// Called when the player has finished their turn.
-        ///// Starts the enemy turn.
-        ///// </summary>
-        //public void EndPlayerTurn()
-        //{
-        //    Debug.Log("Player's turn ended.");
-
-        //    // Reduce cooldowns and update status effects for player
-        //    playerCharacter.GetComponent<Abilities>().ReduceCooldowns();
-        //    playerCharacter.Stats.UpdateStatusEffects();
-        //    // After player turn ends, start enemy turn
-        //    StartAllEnemyTurns();
-        //}
-
         //===============================
         #endregion // ^Turn Management^
 
-        #region Character Positioning and Spawning
+        #region Spawning
         //===============================
 
         private void SpawnEnemies()
@@ -173,7 +137,7 @@ namespace SystemMiami
             for (int i = 0; i < numberOfEnemies; i++)
             {
                 // Find a random unblocked tile to place the enemy
-                OverlayTile spawnTile = GetRandomUnblockedTile();
+                OverlayTile spawnTile = MapManager.MGR.GetRandomUnblockedTile();
 
                 if (spawnTile != null)
                 {
@@ -185,16 +149,14 @@ namespace SystemMiami
                         enemyCombatant = enemyGO.AddComponent<Combatant>();
                     }
 
-                    //// Ensure the enemy has an EnemyController
-                    //EnemyController enemyController = enemyGO.GetComponent<EnemyController>();
-                    //if (enemyController == null)
-                    //{
-                    //    enemyController = enemyGO.AddComponent<EnemyController>();
-                    //    // You can also initialize enemyController properties here if needed
-                    //}
-
                     // Set enemy ID
                     enemyCombatant.ID = i + 1;
+
+                    // Set enemy name
+                    string newName = enemyCombatant.name;
+                    newName = newName.Replace("(Clone)", "");
+                    newName += $"{enemyCombatant.ID}";
+                    enemyCombatant.name = newName;
 
                     // Position enemy on the tile
                     MapManager.MGR.PositionCharacterOnTile(enemyCombatant, spawnTile);
@@ -210,36 +172,7 @@ namespace SystemMiami
                 }
             }
         }
-
-
-        /// <summary>
-        /// Finds a random unblocked tile on the map.
-        /// </summary>
-        /// <returns>An unblocked OverlayTile or null if none are available.</returns>
-        public OverlayTile GetRandomUnblockedTile()
-        {
-            // Get all unblocked tiles
-            List<OverlayTile> unblockedTiles = new List<OverlayTile>();
-
-            foreach (var tile in MapManager.MGR.map.Values)
-            {
-                if (!tile.isBlocked && tile.currentCharacter == null)
-                {
-                    unblockedTiles.Add(tile);
-                }
-            }
-
-            if (unblockedTiles.Count > 0)
-            {
-                // Select a random tile
-                int index = UnityEngine.Random.Range(0, unblockedTiles.Count);
-                return unblockedTiles[index];
-            }
-
-            return null;
-        }
-
         //===============================
-        #endregion // ^Character Positioning and Spawning^
+        #endregion // ^Spawning^
     }
 }
